lgpd_manager/backend/requirements.txt
# Requirements para o LGPD Manager

# Framework Django e REST
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
markdown==3.5

# Autenticação
djangorestframework-simplejwt==5.3.0

# Utilitários
python-dotenv==1.0.0
pyyaml==6.0.1

# Integração com IA
openai==1.3.0

# Geração de documentos
reportlab==4.0.7  # Para geração de PDFs
jinja2==3.1.2     # Templates para documentos

# Logging e monitoramento
sentry-sdk==1.32.0  # Opcional para monitoramento em produção